+++ pkg/providers/v1/aws.go
@@ -33,13 +33,11 @@ import (
 	"github.com/aws/aws-sdk-go/aws/awserr"
 	"github.com/aws/aws-sdk-go/aws/credentials"
 	"github.com/aws/aws-sdk-go/aws/credentials/stscreds"
-	"github.com/aws/aws-sdk-go/aws/ec2metadata"
 	"github.com/aws/aws-sdk-go/aws/endpoints"
 	"github.com/aws/aws-sdk-go/aws/request"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/autoscaling"
 	"github.com/aws/aws-sdk-go/service/ec2"
-	"github.com/aws/aws-sdk-go/service/ec2/ec2iface"
 	"github.com/aws/aws-sdk-go/service/elb"
 	"github.com/aws/aws-sdk-go/service/elbv2"
 	"github.com/aws/aws-sdk-go/service/kms"
@@ -57,7 +55,6 @@ import (
 	clientset "k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/kubernetes/scheme"
 	v1core "k8s.io/client-go/kubernetes/typed/core/v1"
-	"k8s.io/client-go/pkg/version"
 	"k8s.io/client-go/tools/cache"
 	"k8s.io/client-go/tools/record"
 	cloudprovider "k8s.io/cloud-provider"
@@ -761,90 +758,9 @@ func (cfg *CloudConfig) getResolver() endpoints.ResolverFunc {
 	}
 }
 
-// awsSdkEC2 is an implementation of the EC2 interface, backed by aws-sdk-go
-type awsSdkEC2 struct {
-	ec2 ec2iface.EC2API
-}
-
 // Interface to make the CloudConfig immutable for awsSDKProvider
 type awsCloudConfigProvider interface {
 	getResolver() endpoints.ResolverFunc
-}
-
-type awsSDKProvider struct {
-	creds *credentials.Credentials
-	cfg   awsCloudConfigProvider
-
-	mutex          sync.Mutex
-	regionDelayers map[string]*CrossRequestRetryDelay
-}
-
-func newAWSSDKProvider(creds *credentials.Credentials, cfg *CloudConfig) *awsSDKProvider {
-	return &awsSDKProvider{
-		creds:          creds,
-		cfg:            cfg,
-		regionDelayers: make(map[string]*CrossRequestRetryDelay),
-	}
-}
-
-func (p *awsSDKProvider) addHandlers(regionName string, h *request.Handlers) {
-	h.Build.PushFrontNamed(request.NamedHandler{
-		Name: "k8s/user-agent",
-		Fn:   request.MakeAddToUserAgentHandler("kubernetes", version.Get().String()),
-	})
-
-	h.Sign.PushFrontNamed(request.NamedHandler{
-		Name: "k8s/logger",
-		Fn:   awsHandlerLogger,
-	})
-
-	delayer := p.getCrossRequestRetryDelay(regionName)
-	if delayer != nil {
-		h.Sign.PushFrontNamed(request.NamedHandler{
-			Name: "k8s/delay-presign",
-			Fn:   delayer.BeforeSign,
-		})
-
-		h.AfterRetry.PushFrontNamed(request.NamedHandler{
-			Name: "k8s/delay-afterretry",
-			Fn:   delayer.AfterRetry,
-		})
-	}
-
-	p.addAPILoggingHandlers(h)
-}
-
-func (p *awsSDKProvider) addAPILoggingHandlers(h *request.Handlers) {
-	h.Send.PushBackNamed(request.NamedHandler{
-		Name: "k8s/api-request",
-		Fn:   awsSendHandlerLogger,
-	})
-
-	h.ValidateResponse.PushFrontNamed(request.NamedHandler{
-		Name: "k8s/api-validate-response",
-		Fn:   awsValidateResponseHandlerLogger,
-	})
-}
-
-// Get a CrossRequestRetryDelay, scoped to the region, not to the request.
-// This means that when we hit a limit on a call, we will delay _all_ calls to the API.
-// We do this to protect the AWS account from becoming overloaded and effectively locked.
-// We also log when we hit request limits.
-// Note that this delays the current goroutine; this is bad behaviour and will
-// likely cause k8s to become slow or unresponsive for cloud operations.
-// However, this throttle is intended only as a last resort.  When we observe
-// this throttling, we need to address the root cause (e.g. add a delay to a
-// controller retry loop)
-func (p *awsSDKProvider) getCrossRequestRetryDelay(regionName string) *CrossRequestRetryDelay {
-	p.mutex.Lock()
-	defer p.mutex.Unlock()
-
-	delayer, found := p.regionDelayers[regionName]
-	if !found {
-		delayer = NewCrossRequestRetryDelay()
-		p.regionDelayers[regionName] = delayer
-	}
-	return delayer
 }
 
 // InstanceIDIndexFunc indexes based on a Node's instance ID found in its spec.providerID
@@ -877,126 +793,6 @@ func (c *Cloud) SetInformers(informerFactory informers.SharedInformerFactory) {
 	})
 }
 
-func (p *awsSDKProvider) Compute(regionName string) (EC2, error) {
-	awsConfig := &aws.Config{
-		Region:      &regionName,
-		Credentials: p.creds,
-	}
-	awsConfig = awsConfig.WithCredentialsChainVerboseErrors(true).
-		WithEndpointResolver(p.cfg.getResolver())
-	sess, err := session.NewSessionWithOptions(session.Options{
-		Config:            *awsConfig,
-		SharedConfigState: session.SharedConfigEnable,
-	})
-
-	if err != nil {
-		return nil, fmt.Errorf("unable to initialize AWS session: %v", err)
-	}
-	service := ec2.New(sess)
-
-	p.addHandlers(regionName, &service.Handlers)
-
-	ec2 := &awsSdkEC2{
-		ec2: service,
-	}
-	return ec2, nil
-}
-
-func (p *awsSDKProvider) LoadBalancing(regionName string) (ELB, error) {
-	awsConfig := &aws.Config{
-		Region:      &regionName,
-		Credentials: p.creds,
-	}
-	awsConfig = awsConfig.WithCredentialsChainVerboseErrors(true).
-		WithEndpointResolver(p.cfg.getResolver())
-	sess, err := session.NewSessionWithOptions(session.Options{
-		Config:            *awsConfig,
-		SharedConfigState: session.SharedConfigEnable,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("unable to initialize AWS session: %v", err)
-	}
-	elbClient := elb.New(sess)
-	p.addHandlers(regionName, &elbClient.Handlers)
-
-	return elbClient, nil
-}
-
-func (p *awsSDKProvider) LoadBalancingV2(regionName string) (ELBV2, error) {
-	awsConfig := &aws.Config{
-		Region:      &regionName,
-		Credentials: p.creds,
-	}
-	awsConfig = awsConfig.WithCredentialsChainVerboseErrors(true).
-		WithEndpointResolver(p.cfg.getResolver())
-	sess, err := session.NewSessionWithOptions(session.Options{
-		Config:            *awsConfig,
-		SharedConfigState: session.SharedConfigEnable,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("unable to initialize AWS session: %v", err)
-	}
-	elbClient := elbv2.New(sess)
-
-	p.addHandlers(regionName, &elbClient.Handlers)
-
-	return elbClient, nil
-}
-
-func (p *awsSDKProvider) Autoscaling(regionName string) (ASG, error) {
-	awsConfig := &aws.Config{
-		Region:      &regionName,
-		Credentials: p.creds,
-	}
-	awsConfig = awsConfig.WithCredentialsChainVerboseErrors(true).
-		WithEndpointResolver(p.cfg.getResolver())
-	sess, err := session.NewSessionWithOptions(session.Options{
-		Config:            *awsConfig,
-		SharedConfigState: session.SharedConfigEnable,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("unable to initialize AWS session: %v", err)
-	}
-	client := autoscaling.New(sess)
-
-	p.addHandlers(regionName, &client.Handlers)
-
-	return client, nil
-}
-
-func (p *awsSDKProvider) Metadata() (EC2Metadata, error) {
-	sess, err := session.NewSession(&aws.Config{
-		EndpointResolver: p.cfg.getResolver(),
-	})
-	if err != nil {
-		return nil, fmt.Errorf("unable to initialize AWS session: %v", err)
-	}
-	client := ec2metadata.New(sess)
-	p.addAPILoggingHandlers(&client.Handlers)
-	return client, nil
-}
-
-func (p *awsSDKProvider) KeyManagement(regionName string) (KMS, error) {
-	awsConfig := &aws.Config{
-		Region:      &regionName,
-		Credentials: p.creds,
-	}
-	awsConfig = awsConfig.WithCredentialsChainVerboseErrors(true).
-		WithEndpointResolver(p.cfg.getResolver())
-	sess, err := session.NewSessionWithOptions(session.Options{
-		Config:            *awsConfig,
-		SharedConfigState: session.SharedConfigEnable,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("unable to initialize AWS session: %v", err)
-	}
-	kmsClient := kms.New(sess)
-
-	p.addHandlers(regionName, &kmsClient.Handlers)
-
-	return kmsClient, nil
-}
-
 func newEc2Filter(name string, values ...string) *ec2.Filter {
 	filter := &ec2.Filter{
 		Name: aws.String(name),
@@ -1015,254 +811,6 @@ func (c *Cloud) AddSSHKeyToAllInstances(ctx context.Context, user string, keyDat
 // CurrentNodeName returns the name of the current node
 func (c *Cloud) CurrentNodeName(ctx context.Context, hostname string) (types.NodeName, error) {
 	return c.selfAWSInstance.nodeName, nil
-}
-
-// Implementation of EC2.Instances
-func (s *awsSdkEC2) DescribeInstances(request *ec2.DescribeInstancesInput) ([]*ec2.Instance, error) {
-	// Instances are paged
-	results := []*ec2.Instance{}
-	var nextToken *string
-	requestTime := time.Now()
-
-	if request.MaxResults == nil && len(request.InstanceIds) == 0 {
-		// MaxResults must be set in order for pagination to work
-		// MaxResults cannot be set with InstanceIds
-		request.MaxResults = aws.Int64(1000)
-	}
-
-	for {
-		response, err := s.ec2.DescribeInstances(request)
-		if err != nil {
-			recordAWSMetric("describe_instance", 0, err)
-			return nil, fmt.Errorf("error listing AWS instances: %q", err)
-		}
-
-		for _, reservation := range response.Reservations {
-			results = append(results, reservation.Instances...)
-		}
-
-		nextToken = response.NextToken
-		if aws.StringValue(nextToken) == "" {
-			break
-		}
-		request.NextToken = nextToken
-	}
-	timeTaken := time.Since(requestTime).Seconds()
-	recordAWSMetric("describe_instance", timeTaken, nil)
-	return results, nil
-}
-
-// DescribeNetworkInterfaces describes network interface provided in the input.
-func (s *awsSdkEC2) DescribeNetworkInterfaces(input *ec2.DescribeNetworkInterfacesInput) (*ec2.DescribeNetworkInterfacesOutput, error) {
-	requestTime := time.Now()
-	resp, err := s.ec2.DescribeNetworkInterfaces(input)
-	timeTaken := time.Since(requestTime).Seconds()
-	recordAWSMetric("describe_network_interfaces", timeTaken, err)
-	return resp, err
-}
-
-// Implements EC2.DescribeSecurityGroups
-func (s *awsSdkEC2) DescribeSecurityGroups(request *ec2.DescribeSecurityGroupsInput) ([]*ec2.SecurityGroup, error) {
-	// Security groups are paged
-	results := []*ec2.SecurityGroup{}
-	var nextToken *string
-	requestTime := time.Now()
-	for {
-		response, err := s.ec2.DescribeSecurityGroups(request)
-		if err != nil {
-			recordAWSMetric("describe_security_groups", 0, err)
-			return nil, fmt.Errorf("error listing AWS security groups: %q", err)
-		}
-
-		results = append(results, response.SecurityGroups...)
-
-		nextToken = response.NextToken
-		if aws.StringValue(nextToken) == "" {
-			break
-		}
-		request.NextToken = nextToken
-	}
-	timeTaken := time.Since(requestTime).Seconds()
-	recordAWSMetric("describe_security_groups", timeTaken, nil)
-	return results, nil
-}
-
-func (s *awsSdkEC2) AttachVolume(request *ec2.AttachVolumeInput) (*ec2.VolumeAttachment, error) {
-	requestTime := time.Now()
-	resp, err := s.ec2.AttachVolume(request)
-	timeTaken := time.Since(requestTime).Seconds()
-	recordAWSMetric("attach_volume", timeTaken, err)
-	return resp, err
-}
-
-func (s *awsSdkEC2) DetachVolume(request *ec2.DetachVolumeInput) (*ec2.VolumeAttachment, error) {
-	requestTime := time.Now()
-	resp, err := s.ec2.DetachVolume(request)
-	timeTaken := time.Since(requestTime).Seconds()
-	recordAWSMetric("detach_volume", timeTaken, err)
-	return resp, err
-}
-
-func (s *awsSdkEC2) DescribeVolumes(request *ec2.DescribeVolumesInput) ([]*ec2.Volume, error) {
-	// Volumes are paged
-	results := []*ec2.Volume{}
-	var nextToken *string
-	requestTime := time.Now()
-	for {
-		response, err := s.ec2.DescribeVolumes(request)
-
-		if err != nil {
-			recordAWSMetric("describe_volume", 0, err)
-			return nil, err
-		}
-
-		results = append(results, response.Volumes...)
-
-		nextToken = response.NextToken
-		if aws.StringValue(nextToken) == "" {
-			break
-		}
-		request.NextToken = nextToken
-	}
-	timeTaken := time.Since(requestTime).Seconds()
-	recordAWSMetric("describe_volume", timeTaken, nil)
-	return results, nil
-}
-
-func (s *awsSdkEC2) CreateVolume(request *ec2.CreateVolumeInput) (*ec2.Volume, error) {
-	requestTime := time.Now()
-	resp, err := s.ec2.CreateVolume(request)
-	timeTaken := time.Since(requestTime).Seconds()
-	recordAWSMetric("create_volume", timeTaken, err)
-	return resp, err
-}
-
-func (s *awsSdkEC2) DeleteVolume(request *ec2.DeleteVolumeInput) (*ec2.DeleteVolumeOutput, error) {
-	requestTime := time.Now()
-	resp, err := s.ec2.DeleteVolume(request)
-	timeTaken := time.Since(requestTime).Seconds()
-	recordAWSMetric("delete_volume", timeTaken, err)
-	return resp, err
-}
-
-func (s *awsSdkEC2) ModifyVolume(request *ec2.ModifyVolumeInput) (*ec2.ModifyVolumeOutput, error) {
-	requestTime := time.Now()
-	resp, err := s.ec2.ModifyVolume(request)
-	timeTaken := time.Since(requestTime).Seconds()
-	recordAWSMetric("modify_volume", timeTaken, err)
-	return resp, err
-}
-
-func (s *awsSdkEC2) DescribeVolumeModifications(request *ec2.DescribeVolumesModificationsInput) ([]*ec2.VolumeModification, error) {
-	requestTime := time.Now()
-	results := []*ec2.VolumeModification{}
-	var nextToken *string
-	for {
-		resp, err := s.ec2.DescribeVolumesModifications(request)
-		if err != nil {
-			recordAWSMetric("describe_volume_modification", 0, err)
-			return nil, fmt.Errorf("error listing volume modifictions : %v", err)
-		}
-		results = append(results, resp.VolumesModifications...)
-		nextToken = resp.NextToken
-		if aws.StringValue(nextToken) == "" {
-			break
-		}
-		request.NextToken = nextToken
-	}
-	timeTaken := time.Since(requestTime).Seconds()
-	recordAWSMetric("describe_volume_modification", timeTaken, nil)
-	return results, nil
-}
-
-func (s *awsSdkEC2) DescribeSubnets(request *ec2.DescribeSubnetsInput) ([]*ec2.Subnet, error) {
-	// Subnets are not paged
-	response, err := s.ec2.DescribeSubnets(request)
-	if err != nil {
-		return nil, fmt.Errorf("error listing AWS subnets: %q", err)
-	}
-	return response.Subnets, nil
-}
-
-func (s *awsSdkEC2) DescribeAvailabilityZones(request *ec2.DescribeAvailabilityZonesInput) ([]*ec2.AvailabilityZone, error) {
-	// AZs are not paged
-	response, err := s.ec2.DescribeAvailabilityZones(request)
-	if err != nil {
-		return nil, fmt.Errorf("error listing AWS availability zones: %q", err)
-	}
-	return response.AvailabilityZones, err
-}
-
-func (s *awsSdkEC2) CreateSecurityGroup(request *ec2.CreateSecurityGroupInput) (*ec2.CreateSecurityGroupOutput, error) {
-	return s.ec2.CreateSecurityGroup(request)
-}
-
-func (s *awsSdkEC2) DeleteSecurityGroup(request *ec2.DeleteSecurityGroupInput) (*ec2.DeleteSecurityGroupOutput, error) {
-	return s.ec2.DeleteSecurityGroup(request)
-}
-
-func (s *awsSdkEC2) AuthorizeSecurityGroupIngress(request *ec2.AuthorizeSecurityGroupIngressInput) (*ec2.AuthorizeSecurityGroupIngressOutput, error) {
-	return s.ec2.AuthorizeSecurityGroupIngress(request)
-}
-
-func (s *awsSdkEC2) RevokeSecurityGroupIngress(request *ec2.RevokeSecurityGroupIngressInput) (*ec2.RevokeSecurityGroupIngressOutput, error) {
-	return s.ec2.RevokeSecurityGroupIngress(request)
-}
-
-func (s *awsSdkEC2) CreateTags(request *ec2.CreateTagsInput) (*ec2.CreateTagsOutput, error) {
-	requestTime := time.Now()
-	resp, err := s.ec2.CreateTags(request)
-	timeTaken := time.Since(requestTime).Seconds()
-	recordAWSMetric("create_tags", timeTaken, err)
-	return resp, err
-}
-
-func (s *awsSdkEC2) DeleteTags(request *ec2.DeleteTagsInput) (*ec2.DeleteTagsOutput, error) {
-	requestTime := time.Now()
-	resp, err := s.ec2.DeleteTags(request)
-	timeTaken := time.Since(requestTime).Seconds()
-	recordAWSMetric("delete_tags", timeTaken, err)
-	return resp, err
-}
-
-func (s *awsSdkEC2) DescribeRouteTables(request *ec2.DescribeRouteTablesInput) ([]*ec2.RouteTable, error) {
-	results := []*ec2.RouteTable{}
-	var nextToken *string
-	requestTime := time.Now()
-	for {
-		response, err := s.ec2.DescribeRouteTables(request)
-		if err != nil {
-			recordAWSMetric("describe_route_tables", 0, err)
-			return nil, fmt.Errorf("error listing AWS route tables: %q", err)
-		}
-
-		results = append(results, response.RouteTables...)
-
-		nextToken = response.NextToken
-		if aws.StringValue(nextToken) == "" {
-			break
-		}
-		request.NextToken = nextToken
-	}
-	timeTaken := time.Since(requestTime).Seconds()
-	recordAWSMetric("describe_route_tables", timeTaken, nil)
-	return results, nil
-}
-
-func (s *awsSdkEC2) CreateRoute(request *ec2.CreateRouteInput) (*ec2.CreateRouteOutput, error) {
-	return s.ec2.CreateRoute(request)
-}
-
-func (s *awsSdkEC2) DeleteRoute(request *ec2.DeleteRouteInput) (*ec2.DeleteRouteOutput, error) {
-	return s.ec2.DeleteRoute(request)
-}
-
-func (s *awsSdkEC2) ModifyInstanceAttribute(request *ec2.ModifyInstanceAttributeInput) (*ec2.ModifyInstanceAttributeOutput, error) {
-	return s.ec2.ModifyInstanceAttribute(request)
-}
-
-func (s *awsSdkEC2) DescribeVpcs(request *ec2.DescribeVpcsInput) (*ec2.DescribeVpcsOutput, error) {
-	return s.ec2.DescribeVpcs(request)
 }
 
 func init() {
@@ -1970,52 +1518,6 @@ func IsAWSErrorInstanceNotFound(err error) bool {
 // This should be stored as a single letter (i.e. c, not sdc or /dev/sdc)
 type mountDevice string
 
-type awsInstance struct {
-	ec2 EC2
-
-	// id in AWS
-	awsID string
-
-	// node name in k8s
-	nodeName types.NodeName
-
-	// availability zone the instance resides in
-	availabilityZone string
-
-	// ID of VPC the instance resides in
-	vpcID string
-
-	// ID of subnet the instance resides in
-	subnetID string
-
-	// instance type
-	instanceType string
-}
-
-// newAWSInstance creates a new awsInstance object
-func newAWSInstance(ec2Service EC2, instance *ec2.Instance) *awsInstance {
-	az := ""
-	if instance.Placement != nil {
-		az = aws.StringValue(instance.Placement.AvailabilityZone)
-	}
-	self := &awsInstance{
-		ec2:              ec2Service,
-		awsID:            aws.StringValue(instance.InstanceId),
-		nodeName:         mapInstanceToNodeName(instance),
-		availabilityZone: az,
-		instanceType:     aws.StringValue(instance.InstanceType),
-		vpcID:            aws.StringValue(instance.VpcId),
-		subnetID:         aws.StringValue(instance.SubnetId),
-	}
-
-	return self
-}
-
-// Gets the full information about this instance from the EC2 API
-func (i *awsInstance) describeInstance() (*ec2.Instance, error) {
-	return describeInstance(i.ec2, InstanceID(i.awsID))
-}
-
 // Gets the mountDevice already assigned to the volume, or assigns an unused mountDevice.
 // If the volume is already assigned, this will return the existing mountDevice with alreadyAttached=true.
 // Otherwise the mountDevice is assigned by finding the first available mountDevice, and it is returned with alreadyAttached=false.
@@ -2123,120 +1625,6 @@ func (c *Cloud) endAttaching(i *awsInstance, volumeID EBSVolumeID, mountDevice m
 	klog.V(2).Infof("Releasing in-process attachment entry: %s -> volume %s", mountDevice, volumeID)
 	delete(c.attaching[i.nodeName], mountDevice)
 	return true
-}
-
-type awsDisk struct {
-	ec2 EC2
-
-	// Name in k8s
-	name KubernetesVolumeID
-	// id in AWS
-	awsID EBSVolumeID
-}
-
-func newAWSDisk(aws *Cloud, name KubernetesVolumeID) (*awsDisk, error) {
-	awsID, err := name.MapToAWSVolumeID()
-	if err != nil {
-		return nil, err
-	}
-	disk := &awsDisk{ec2: aws.ec2, name: name, awsID: awsID}
-	return disk, nil
-}
-
-// Helper function for describeVolume callers. Tries to retype given error to AWS error
-// and returns true in case the AWS error is "InvalidVolume.NotFound", false otherwise
-func isAWSErrorVolumeNotFound(err error) bool {
-	if err != nil {
-		if awsError, ok := err.(awserr.Error); ok {
-			// https://docs.aws.amazon.com/AWSEC2/latest/APIReference/errors-overview.html
-			if awsError.Code() == "InvalidVolume.NotFound" {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-// Gets the full information about this volume from the EC2 API
-func (d *awsDisk) describeVolume() (*ec2.Volume, error) {
-	volumeID := d.awsID
-
-	request := &ec2.DescribeVolumesInput{
-		VolumeIds: []*string{volumeID.awsString()},
-	}
-
-	volumes, err := d.ec2.DescribeVolumes(request)
-	if err != nil {
-		return nil, err
-	}
-	if len(volumes) == 0 {
-		return nil, fmt.Errorf("no volumes found")
-	}
-	if len(volumes) > 1 {
-		return nil, fmt.Errorf("multiple volumes found")
-	}
-	return volumes[0], nil
-}
-
-func (d *awsDisk) describeVolumeModification() (*ec2.VolumeModification, error) {
-	volumeID := d.awsID
-	request := &ec2.DescribeVolumesModificationsInput{
-		VolumeIds: []*string{volumeID.awsString()},
-	}
-	volumeMods, err := d.ec2.DescribeVolumeModifications(request)
-
-	if err != nil {
-		return nil, fmt.Errorf("error describing volume modification %s with %v", volumeID, err)
-	}
-
-	if len(volumeMods) == 0 {
-		return nil, fmt.Errorf("no volume modifications found for %s", volumeID)
-	}
-	lastIndex := len(volumeMods) - 1
-	return volumeMods[lastIndex], nil
-}
-
-func (d *awsDisk) modifyVolume(requestGiB int64) (int64, error) {
-	volumeID := d.awsID
-
-	request := &ec2.ModifyVolumeInput{
-		VolumeId: volumeID.awsString(),
-		Size:     aws.Int64(requestGiB),
-	}
-	output, err := d.ec2.ModifyVolume(request)
-	if err != nil {
-		modifyError := fmt.Errorf("AWS modifyVolume failed for %s with %v", volumeID, err)
-		return requestGiB, modifyError
-	}
-
-	volumeModification := output.VolumeModification
-
-	if aws.StringValue(volumeModification.ModificationState) == ec2.VolumeModificationStateCompleted {
-		return aws.Int64Value(volumeModification.TargetSize), nil
-	}
-
-	backoff := wait.Backoff{
-		Duration: 1 * time.Second,
-		Factor:   2,
-		Steps:    10,
-	}
-
-	checkForResize := func() (bool, error) {
-		volumeModification, err := d.describeVolumeModification()
-
-		if err != nil {
-			return false, err
-		}
-
-		// According to https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/monitoring_mods.html
-		// Size changes usually take a few seconds to complete and take effect after a volume is in the Optimizing state.
-		if aws.StringValue(volumeModification.ModificationState) == ec2.VolumeModificationStateOptimizing {
-			return true, nil
-		}
-		return false, nil
-	}
-	waitWithErr := wait.ExponentialBackoff(backoff, checkForResize)
-	return requestGiB, waitWithErr
 }
 
 // applyUnSchedulableTaint applies a unschedulable taint to a node after verifying
